+++ packages/scxa-cell-type-wheel/src/index.js
@@ -1,1 +1,1 @@
-export { default } from './MyComponent'
+export { default } from './CellTypeWheel'
